gtk: drop unecessary manual code
gdk::Event (and fundamental types) code can now be generated correctly
