+++ gtk4/src/lib.rs
@@ -123,9 +123,7 @@ mod border;
 mod builder;
 mod builder_rust_scope;
 mod cell_area;
-mod cell_editable;
 mod cell_layout;
-mod cell_renderer;
 mod closure_expression;
 mod color_chooser;
 mod combo_box;
@@ -153,7 +151,6 @@ mod font_chooser;
 mod functions;
 mod gesture_stylus;
 mod icon_theme;
-mod im_context;
 mod im_context_simple;
 mod info_bar;
 mod keyval_trigger;
@@ -191,7 +188,6 @@ mod spin_button;
 mod string_list;
 mod text;
 mod text_buffer;
-mod text_view;
 mod tree_model;
 mod tree_model_filter;
 mod tree_path;
