+++ gtk4/src/prelude.rs
@@ -8,9 +8,7 @@ pub use crate::auto::traits::*;
 pub use crate::accessible::AccessibleExtManual;
 pub use crate::actionable::ActionableExtManual;
 pub use crate::cell_area::CellAreaExtManual;
-pub use crate::cell_editable::CellEditableExtManual;
 pub use crate::cell_layout::CellLayoutExtManual;
-pub use crate::cell_renderer::CellRendererExtManual;
 pub use crate::color_chooser::ColorChooserExtManual;
 pub use crate::combo_box::ComboBoxExtManual;
 pub use crate::dialog::DialogExtManual;
@@ -21,13 +19,11 @@ pub use crate::entry_buffer::EntryBufferExtManual;
 pub use crate::expression::{GObjectPropertyExpressionExt, IsExpression};
 pub use crate::file_chooser::FileChooserExtManual;
 pub use crate::font_chooser::FontChooserExtManual;
-pub use crate::im_context::IMContextExtManual;
 pub use crate::media_stream::MediaStreamExtManual;
 pub use crate::native_dialog::NativeDialogExtManual;
 pub use crate::scale::ScaleExtManual;
 pub use crate::shortcut_trigger::ShortcutTriggerExtManual;
 pub use crate::text_buffer::TextBufferExtManual;
-pub use crate::text_view::TextViewExtManual;
 pub use crate::tree_model::TreeModelExtManual;
 pub use crate::tree_model_filter::TreeModelFilterExtManual;
 pub use crate::tree_sortable::TreeSortableExtManual;
